Extract "member" related tests into own file
