+++ tests/browser/support/meta/meta.test.ts
@@ -259,35 +259,7 @@ describe('@meta() decorator', () => {
         // Debug
         // new MyClass();
 
-        expect(Metadata.get(MyClass, ['static', 'methods', 'myMethod', MY_KEY])).toBe(
-            'secret-value',
-        );
+        expect(Metadata.get(MyClass, ['static', 'methods', 'myMethod', MY_KEY]))
+            .toBe('secret-value');
     });
-
-    test('can get meta using instance method reference', () => {
-        class MyService
-        {
-            @meta('foo', 'bar')
-            play()
-            {/* empty */}
-        }
-
-        // Debug
-        const instance = new MyService();
-        const { play } = instance; // eslint-disable-line @typescript-eslint/unbound-method
-
-        expect(Metadata.has(play, 'foo'), 'Unable to determine if has meta for instance method')
-            .toBe(true);
-
-        expect(Metadata.get(play, 'foo'), 'Incorrect meta value for instance method')
-            .toBe('bar');
-    });
-
-    // TODO: Inherit instance method meta, ... obtain via member directly.
-    // TODO: getter / setter (field) meta, ... obtain via member directly.
-    // TODO: Static method meta, ... obtain via static member directly.
-    // TODO: Inherit static method meta, ... obtain via static member directly.
-    // TODO: Static getter / setter (field) meta, ... obtain via member directly.
-
-    
 });
